Api methods with if else validation

+++ index.js
@@ -28,16 +28,23 @@ app.get('/getuser', async (req, res) => {
   // post query
   app.post('/insertuser', async (req, res) => {
     // const querySql = 'INSERT INTO `users` (`id`, `name`, `rollno`, `classname`, `address`, `mobileno`) VALUES (NULL, 'Mansi', '3', '15 th', 'Sangli', '7890876578');';
+    
     const name = req.body.name;
     const rollno = req.body.rollno;
     const classname = req.body.classname
     const address = req.body.address
     const mobileno = req.body.mobileno
+
+    if(name && rollno && classname && address && mobileno){
     const querySql = `INSERT INTO users ( name, rollno, classname, address, mobileno) VALUES ('${name}', '${rollno}', '${classname}', '${address}', '${mobileno}');`;
     console.log(querySql)
     const rows = await connection({ querys: querySql, values: [] });
     console.log(rows)
     res.send(rows)
+    }else{
+      res.send({message:"invalid data"})
+    }
+    
     })
 
 
@@ -47,11 +54,29 @@ app.get('/getuser', async (req, res) => {
       console.log(id)
       const address = req.body.address
       const mobileno = req.body.mobileno
-      // const querySql = `UPDATE users SET address='Kolhapur', mobileno = '99999999' WHERE id = 1;`;
-      const querySql = `UPDATE users SET address='${address}', mobileno = '${mobileno}' WHERE id = ${id};`;
-      const rows = await connection({ querys: querySql, values: [] });
-      console.log(rows)
-      res.send(rows)
+
+      if(id && address && mobileno){
+        // to get specific id from table
+        const userSql = `SELECT * FROM users where id = '${id}'`;
+        const userRows = await connection({ querys: userSql, values: [] });
+        console.log(userRows)
+        // console.log(userRows[0].Roll)    //get array item 
+        // condition for id not empty ,lenght not grater than 0 and roll==user
+        if(userRows && userRows.length > 0 &&  userRows[0].Roll=="User"){
+          const querySql = `UPDATE users SET address='${address}', mobileno = '${mobileno}' WHERE id = ${id};`;
+          const rows = await connection({ querys: querySql, values: [] });
+          console.log(rows)
+          res.send(rows)
+        }
+        else{
+          res.send({message:"Admin not update data"})
+        }
+        
+      }
+      else{
+        res.send({message:"user not update"})
+      }
+     
       })
 
 
@@ -59,10 +84,16 @@ app.get('/getuser', async (req, res) => {
       app.delete('/deleteuser/:id', async (req, res) => {
         const {id} = req.params
         console.log(id)
-        const querySql = `DELETE FROM users WHERE id=${id};`;
-        const rows = await connection({ querys: querySql, values: [] });
-        console.log(rows)
-        res.send(rows)
+
+        if(id){
+          const querySql = `DELETE FROM users WHERE id=${id};`;
+          const rows = await connection({ querys: querySql, values: [] });
+          console.log(rows)
+          res.send(rows)
+        }else{
+          res.send({message:"user deleted"})
+        }
+       
         })
